🔨 Restructure Options.vue
